10/7/2026 : Version 1.0 (Functional but confusing) Lots to improve but tunggu Aasim & Haziq add bahagian diorg dulu

+++ src/main/java/Delivery/Node.java
@@ -1,23 +1,16 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
+// Node.java
 package Delivery;
 
-/**
- *
- * @author ammar
- */
 class Node {
     String foodName;
-    int price;  
+    double price;
+
     Node left, right;
     
-    Node(String foodName, int price){
-    this.foodName = foodName;
-    this.price = price;
-    this.left = null;
-    this.right = null;
+    Node(String foodName, double price) {
+        this.foodName = foodName;
+        this.price = price;
+        this.left = null;
+        this.right = null;
     }
 }
-
